% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLength.R
\name{allowedExons}
\alias{allowedExons}
\title{Get allowedExons number}
\usage{
allowedExons(symbols = NULL, species = c("Hs", "Mm"))
}
\arguments{
\item{symbols}{character vector of HGNC gene symbols}

\item{species}{The species of gene. only support Hs or Mm}
}
\description{
Get allowedExons from \code{OrganismDb} package
}
\note{
Only consider the gene on chr 1-22, chrX and chrY
}
\examples{
\dontrun{
allowedExons('TP53')
allowedExons('Pzp', species = 'Mm')
}

}
