% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_info.R
\name{check_id}
\alias{check_id}
\alias{is_symbol}
\alias{is_entrez}
\alias{is_ensembl}
\alias{is_refseq}
\alias{is_accnum}
\alias{is_ucsc}
\title{Check Gene id}
\usage{
is_symbol(genes, species = c("Hs", "Mm"))

is_entrez(genes, species = c("Hs", "Mm"))

is_ensembl(genes, species = c("Hs", "Mm"))

is_refseq(genes, species = c("Hs", "Mm"))

is_accnum(genes, species = c("Hs", "Mm"))

is_ucsc(genes, species = c("Hs", "Mm"))
}
\arguments{
\item{genes}{gene id to check}

\item{species}{The species of gene. only support Hs or Mm}
}
\value{
return a logical vector
}
\description{
Check Gene id whether is matched with specific id type
}
\examples{
\dontrun{

is_symbol('TP53', species = 'Hs')
is_symbol(c('TP53', 'SJDSJ'))

}

}
