% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_id.R
\name{convert_id}
\alias{convert_id}
\alias{alias2symbol}
\alias{alias2entrez}
\alias{alias2ucsc}
\alias{alias2refseq}
\alias{alias2ensembl}
\alias{symbol2entrez}
\alias{symbol2ensembl}
\alias{symbol2ucsc}
\alias{symbol2refseq}
\alias{symbol2alias}
\alias{entrez2symbol}
\alias{entrez2ucsc}
\alias{entrez2refseq}
\alias{entrez2alias}
\alias{entrez2entrez}
\alias{ucsc2symbol}
\alias{ucsc2refseq}
\alias{ucsc2alias}
\alias{ucsc2entrez}
\alias{ucsc2ensembl}
\alias{refseq2symbol}
\alias{refseq2ucsc}
\alias{refseq2alias}
\alias{refseq2entrez}
\alias{refseq2ensembl}
\alias{ensembl2symbol}
\alias{ensembl2refseq}
\alias{ensembl2ucsc}
\alias{ensembl2alias}
\alias{ensembl2entrez}
\alias{entrez2ensembl}
\title{Search genes annotation in \code{OrganismDb} object}
\usage{
alias2symbol(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

alias2entrez(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

alias2ucsc(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

alias2refseq(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

alias2ensembl(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

symbol2entrez(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

symbol2ensembl(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

symbol2refseq(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

symbol2ucsc(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

symbol2alias(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

entrez2symbol(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

entrez2ensembl(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

entrez2refseq(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

entrez2ucsc(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

entrez2alias(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

refseq2entrez(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

refseq2symbol(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

refseq2ensembl(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

refseq2ucsc(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

refseq2alias(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

ucsc2symbol(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

ucsc2ensembl(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

ucsc2refseq(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

ucsc2entrez(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")

ucsc2alias(keys = NULL, species = c("Hs", "Mm"), multiVals = "first")
}
\arguments{
\item{keys}{Gene id to convert}

\item{species}{The species of gene. only support Hs or Mm}

\item{multiVals}{The same with \code{\link[AnnotationDbi]{mapIds}}}
}
\value{
A data frame with annotation info
}
\description{
Search keytype to more annotation type in \code{OrganismDb} object
}
\examples{
\dontrun{

symbol2entrez("TP53", species = "Hs")

}

}
\seealso{
\code{\link[AnnotationDbi]{ACCNUM}}
\code{\link[Homo.sapiens]{Homo.sapiens}}
\code{\link[Mus.musculus]{Mus.musculus}}
}
