% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLength.R
\name{get_exons_info}
\alias{get_exons_info}
\title{Get Gene exon length information}
\usage{
get_exons_info(symbols = NULL, species = c("Hs", "Mm"))
}
\arguments{
\item{symbols}{character vector of HGNC gene symbols}

\item{species}{The species of gene. only support Hs or Mm}
}
\description{
Get gene info from \code{OrganismDb} package
}
\note{
Only consider the gene on chr 1-22, chrX and chrY
}
\examples{
\dontrun{
get_exons_info('TP53')
get_exons_info('Pzp', species = 'Mm')
}

}
