% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord.R
\name{get_gene_coord_from_ensembl}
\alias{get_gene_coord_from_ensembl}
\title{Get gene information from biomaRt}
\usage{
get_gene_coord_from_ensembl(
  species = c("Hs", "Mm"),
  hg19 = FALSE,
  mm10 = FALSE,
  ...
)
}
\arguments{
\item{species}{choose the species, 'Hs' or 'Mm'}

\item{hg19}{It is specific to \code{species = 'Hs'}, whether download hg109 information}

\item{mm10}{It is specific to \code{species = 'Mm'}, whether download mm10 information}

\item{...}{Other argument of \code{\link[biomaRt]{useEnsembl}}}
}
\value{
return a data.frame object of gene coordinate information
}
\description{
Get  information from biomaRt
}
\seealso{
https://www.ensembl.org/info/website/archives/assembly.html
}
