% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord.R
\name{get_gene_coord_from_txdb}
\alias{get_gene_coord_from_txdb}
\title{Get gene information from bioconductor}
\usage{
get_gene_coord_from_txdb(species = c("Hs", "Mm"), hg19 = FALSE, mm10 = FALSE)
}
\arguments{
\item{species}{choose the species, 'Hs' or 'Mm'}

\item{hg19}{It is specific to \code{species = 'Hs'}, whether download hg109 information}

\item{mm10}{It is specific to \code{species = 'Mm'}, whether download mm10 information}
}
\value{
return a data.frame object of gene coordinate information
}
\description{
Get  information from bioconductor \code{db} package and corresponding \code{Txdb} package
}
\seealso{
\code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
\code{\link[TxDb.Hsapiens.UCSC.hg19.knownGene]{TxDb.Hsapiens.UCSC.hg19.knownGene}}
\code{\link[TxDb.Hsapiens.UCSC.hg38.knownGene]{TxDb.Hsapiens.UCSC.hg38.knownGene}}
\code{\link[org.Mm.eg.db]{org.Mm.eg.db}}
\code{\link[TxDb.Mmusculus.UCSC.mm10.knownGene]{TxDb.Mmusculus.UCSC.mm10.knownGene}}
\code{\link[TxDb.Mmusculus.UCSC.mm39.knownGene]{TxDb.Mmusculus.UCSC.mm39.knownGene}}
}
