% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneInfo.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Get Genomic Information}
\usage{
get_gene_info(symbols, species = c("Hs", "Mm"))
}
\arguments{
\item{symbols}{character vector of HGNC gene symbols}

\item{species}{The species of gene. only support Hs or Mm}
}
\description{
Get gene info from bioconductor \code{db} package and corresponding \code{Txdb} package
}
\examples{
\dontrun{
get_gene_info('TP53')
get_gene_info('Pzp', species = 'Mm')
}

}
