% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_info.R
\name{get_geneid_type}
\alias{get_geneid_type}
\title{Check gene id type}
\usage{
get_geneid_type(
  symbol,
  species = c("Hs", "Mm"),
  aliasesBeforeSymbols = FALSE,
  priority = NULL
)
}
\arguments{
\item{symbol}{The gene symbol to check}

\item{species}{The species of gene. only support Hs or Mm}

\item{aliasesBeforeSymbols}{check aliases before Symbols. If TRUE, when aliases is equal to Symbol, will return 'aliases'. Default FALSE.}

\item{priority}{The priority to check gene id type. you can set a vector of id type.}
}
\value{
return a vector with symbol type
}
\description{
Check gene id type by specific gene
}
\examples{
\dontrun{

get_geneid_type('TP53', species = 'Hs')
get_geneid_type(c('TP53', 'ADSL'), species = 'Hs')

}


}
\seealso{
\code{\link[AnnotationDbi]{ACCNUM}}
\code{\link[Homo.sapiens]{Homo.sapiens}}
\code{\link[Mus.musculus]{Mus.musculus}}
}
