% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{search_genes}
\alias{search_genes}
\alias{search_entrez}
\alias{search_symbol}
\alias{search_ensembl}
\alias{search_refseq}
\alias{search_alias}
\title{Search genes annotation in \code{OrganismDb} object}
\usage{
search_symbol(
  keys = NULL,
  species = c("Hs", "Mm"),
  columns = NULL,
  add.columns = NULL,
  remove.columns = NULL
)

search_ensembl(
  keys = NULL,
  species = c("Hs", "Mm"),
  columns = NULL,
  add.columns = NULL,
  remove.columns = NULL
)

search_entrez(
  keys = NULL,
  species = c("Hs", "Mm"),
  columns = NULL,
  add.columns = NULL,
  remove.columns = NULL
)

search_refseq(
  keys = NULL,
  species = c("Hs", "Mm"),
  columns = NULL,
  add.columns = NULL,
  remove.columns = NULL
)

search_alias(
  keys = NULL,
  species = c("Hs", "Mm"),
  columns = NULL,
  add.columns = NULL,
  remove.columns = NULL
)
}
\arguments{
\item{keys}{Gene id to search}

\item{species}{The species of gene. only support Hs or Mm}

\item{columns}{The column to return. If NULL, return three column: SYMBOL, ENTREZID, and TXCHROM}

\item{add.columns}{The column to add in the result, the column must in the \code{OrganismDb} object}

\item{remove.columns}{The column to remove in the result, the column must in the result.}
}
\value{
A data frame with annotation info
}
\description{
Search keytype to more annotation type in \code{OrganismDb} object
}
\examples{
\dontrun{

search_symbol('TP53', species = 'Hs')

}


}
\seealso{
\code{\link[AnnotationDbi]{ACCNUM}}
\code{\link[Homo.sapiens]{Homo.sapiens}}
\code{\link[Mus.musculus]{Mus.musculus}}
}
