% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLength.R
\name{totalExons}
\alias{totalExons}
\title{Get totalExons number}
\usage{
totalExons(symbols = NULL, species = c("Hs", "Mm"))
}
\arguments{
\item{symbols}{character vector of HGNC gene symbols}

\item{species}{The species of gene. only support Hs or Mm}
}
\description{
Get totalExons from \code{OrganismDb} package
}
\note{
Only consider the gene on chr 1-22, chrX and chrY
}
\examples{
\dontrun{
totalExons('TP53')
totalExons('Pzp', species = 'Mm')
}

}
